+++ app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Header from "./components/home/headers/header";
 
 const siteUrl =
   process.env.NEXT_PUBLIC_SITE ??
