--- app/login/layout.tsx
@@ -1,16 +1,0 @@
-export const metadata = {
-  title: "Next.js",
-  description: "Generated by Next.js",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      <main>{children}</main>
-    </div>
-  );
-}
